dird: move SaveTree to where it makes more sense
